+++ R/looney_pyramids.R
@@ -61,6 +61,26 @@ makeContent.looney_pyramid_face <- function(x) { # nolint
     setChildren(x, gl)
 }
 
+# Copied over from {oblicubes}
+#' @importFrom grDevices col2rgb
+cheap_darken <- function(col, amount) {
+    mat <- col2rgb(col, alpha = TRUE)
+    mat[1:3, ] <- mat[1:3, ] * (1 - amount)
+    rgb(mat[1, ], mat[2, ], mat[3, ], mat[4, ], maxColorValue = 255)
+}
+
+# Copied over from {gridpattern}
+#' @importFrom grDevices col2rgb rgb
+update_alpha <- function (colour, alpha = NA_real_) {
+    n <- max(lengths(list(colour, alpha)))
+    colour <- rep_len(colour, n)
+    alpha <- rep_len(alpha, n)
+    m <- grDevices::col2rgb(colour, alpha = TRUE)/255
+    m[4, ] <- ifelse(is.na(alpha), m[4, ], alpha)
+    apply(m, 2, function(x) grDevices::rgb(x[1], x[2], x[3],
+        x[4]))
+}
+
 # Pyramid Base 9/16", 25/32", 1"
 # Pyramid Height 1, 1 3/8", 1 3/4"
 # Pyramid Face Height = sqrt((Pyramid Height)^2 + (Base / 2)^2)
@@ -74,20 +94,46 @@ makeContent.looney_pyramid_face <- function(x) { # nolint
 #' See \url{https://www.looneylabs.com/looney-pyramids}.
 #'
 #' @return Returns a list with an \code{icehouse_pieces} configuration for drawing Looney Pyramids.
+#' @param ... Should be empty.
+#' @param border If `TRUE` draw a black border line on piece edges.  Should normally be `TRUE` when
+#'               drawing with `{grid}` graphics and `FALSE` when drawing with 3D graphic systems.
 #' @export
-looney_pyramids <- function() {
-    icehouse_pieces <- list(n_ranks = 3, n_suits = 6,
-                            width.r1.pyramid = 0.5625, width.r2.pyramid = 0.78125, width.r3.pyramid = 1,
-                            height.r1.pyramid = 1.03879813366216633419, height.r2.pyramid = 1.429409979895551074947,
-                            height.r3.pyramid = 1.820027472320129513506,
-                            rank_text = ",\u25cf,\u25cf\u25cf,\u25cf\u25cf\u25cf",
-                            # suit_color = "#D55E00,#808080,#009E73,#56B4E9,#E69F00,#808080",
-                            suit_color = "#B53E00,#808080,#007E53,#3694C9,#C67F00,#808080",
-                            border_color.pyramid = "#D55E00A0,#808080FF,#009E73A0,#56B4E9A0,#E69F00A0,#808080FF",
-                            background_color.pyramid = "#D55E0090,#000000,#009E7390,#56B4E890,#E69F0090,#FFFFFF",
-                            border_lex.pyramid = 4, grob_fn.pyramid = looneyPyramidGrob
-                            )
-    list(icehouse_pieces = pp_cfg(icehouse_pieces))
+looney_pyramids <- function(..., border = TRUE) {
+    cb_suit_colors_pure <- c("#D55E00", "#000000", "#009E73", "#0072B2", "#F0E442", "#FFFFFF")
+    cb_suit_colors_impure <- cb_suit_colors_pure
+    cb_suit_colors_impure[2L] <- "grey30"
+    # suit_color = "#B53E00,#808080,#007E53,#3694C9,#C67F00,#808080"
+    # background_color.pyramid = "#D55E0090,#000000,#009E7390,#56B4E890,#E69F0090,#FFFFFF"
+    # border_color.pyramid = "#D55E00A0,#808080FF,#009E73A0,#56B4E9A0,#E69F00A0,#808080FF"
+    suit_color <- cheap_darken(cb_suit_colors_impure, 0.4)
+    suit_color[2L] <- "black"
+    # suit_color[5:6] <- cheap_darken(cb_suit_colors_impure[5:6], 0.36)
+    suit_color <- update_alpha(suit_color, 1)
+    background_color <- cb_suit_colors_impure
+    background_color[c(1, 3:5)] <- update_alpha(cb_suit_colors_impure[c(1, 3:5)], 0.80)
+    if (border) {
+        colors <- list(suit_color = suit_color,
+                       background_color.pyramid = background_color,
+                       border_color = "black",
+                       border_lex = 4)
+    } else {
+        colors <- list(suit_color = suit_color,
+                       background_color.pyramid = background_color,
+                       border_color = "transparent",
+                       border_lex = 0)
+    }
+    rank_text <- "\u25cf,\u25cf\u25cf,\u25cf\u25cf\u25cf"
+    pyramids <- list(n_ranks = 3, n_suits = 6,
+                     width.r1.pyramid = 0.5625, width.r2.pyramid = 0.78125, width.r3.pyramid = 1,
+                     height.r1.pyramid = 1.03879813366216633419, height.r2.pyramid = 1.429409979895551074947,
+                     height.r3.pyramid = 1.820027472320129513506,
+                     rank_text = rank_text,
+                     grob_fn.pyramid = looneyPyramidGrob)
+    check_dots_empty()
+    icehouse_pieces <- pp_cfg(c(pyramids, colors))
+    icehouse_pieces$has_piecepack <- FALSE
+    icehouse_pieces$has_pyramids <- TRUE
+    list(icehouse_pieces = icehouse_pieces)
 }
 
 # Internal function copied from `{piecepackr}`